Moved voucher.ts out of the infra/ folder
